test: split retry stage runtime tests
